refactored "core" parsers
