+++ imap-proto/src/lib.rs
@@ -4,6 +4,7 @@ extern crate nom;
 pub mod builders;
 mod parser;
 pub mod types;
+pub mod core;
 
 pub use parser::{parse_response, ParseResult};
 pub use types::*;
